[build] Add several util units (cleanup)
