+++ src/node/transaction.h
@@ -8,20 +8,7 @@
 #include <attributes.h>
 #include <primitives/transaction.h>
 #include <uint256.h>
-
-enum class TransactionError {
-    OK, //!< No error
-    MISSING_INPUTS,
-    ALREADY_IN_CHAIN,
-    P2P_DISABLED,
-    MEMPOOL_REJECTED,
-    MEMPOOL_ERROR,
-    INVALID_PSBT,
-    PSBT_MISMATCH,
-    SIGHASH_MISMATCH,
-};
-
-std::string TransactionErrorString(const TransactionError error);
+#include <util/error.h>
 
 /**
  * Broadcast a transaction
